Render HTML DRF fields properly for preferences
Map preference types to proper DRF fields rather than generic 'Field'.

This alone is NOT enough to make the form work as we are currently using
to_db instead of to_python. So the HTML form will send something like
value="true" for a boolean field and currently we don't try to parse
this into a boolean. But this makes the fields render properly, so it's
a start.

Signed-off-by: Rick Elrod <rick@elrod.me>