Use validate() and to_python for preferences
Switch to using (and ensuring we call) validate() on preference updates
instead of relying on to_db to do validation.

This seems to make more logical sense, as when the user gives an invalid
value, it's something that doesn't pass validation (so ValidationError
makes sense rather than SerializationError which seems like it would be
more for cases when we have something (maybe valid) but the serializer
can't convert it into a reasonable value (python or db)).

Using to_python in the serializer allows the DRF HTML form to work for
preferences now, so things like checkboxes for booleans should take
effect as of this change rather than always getting marked as a wrong
type.

Signed-off-by: Rick Elrod <rick@elrod.me>