Allow JWT key to be changed via API
Rather than hardcoding the JWT key in settings, now it can be changed
via the API.

When it changes, the public key (a read-only field) also gets updated
after extracting out the public key from the new private key.

For right now, a blank private key ("") is acceptable. This is because
there is no story around how the installer is going to work and if there
is going to be a way to feed in an initial key, or what.

Signed-off-by: Rick Elrod <rick@elrod.me>