Make settings PUT show all updated section fields
With the addition of on_update in
fba0f747710f883c8512ce33f5fcf23c54235abb, on_update callbacks can update
other preference fields. However, the PUT response would show the old
values instead of the new values after on_update ran.

Instead, fetch the whole section again and show the current values.

In theory I think there could be a small race here if two people update
a section at the exact same time. Since we're not showing the actual PUT
values (since we can't always) and we're doing a fetch of the whole
section, if an update happened between when we updated and when we
fetched, then we could show values the person might not expect. I expect
this to be pretty rare though.

Signed-off-by: Red Hat Contributor <noreply@redhat.com>