Allow django-ansible-base to be edited at runtime (#121)
In the development environment, this allows django-ansible-base to be
edited at runtime. The cost of doing this is a few assumptions imposed
on the developer - namely that there will be a clone of
django-ansible-base as a subdirectory of aap-gateway.

So the expected tree looks like this:

|- aap-gateway
   - aap_gateway_api
   - ...
   - django-ansible-base

During image build time, we temporarily COPY in django-ansible-base.
This allows us to install dependencies and set up the editable install.
This directory is deleted after the pip install is done and at image
runtime, the local directory is mounted in.

If during buildtime a local clone is not found, then
django-ansible-base as installed from git will be used.

Testing infrastructure (tox) has been updated. Tox will install
django-ansible-base from the local clone if it exists, otherwise,
it, too, will install from git.

Signed-off-by: Rick Elrod <rick@elrod.me>