Create csrf generalized test
Tests most endpoints in gateway's django application for blindspots in CSRF
verification, with the motivation of blocking users who introduce holes in
verification when SessionAuthentication is used.

Does not check for CSRF verification for all authenticators, and ignores
endpoints that do not support the OPTIONS method.

If this test introduces a false-positive, it should be xfail-ed.