Allow logout with GET (#599)
The "get" functionality is needed to render a logout link from
component DRF pages.

I'll note that allowing GET on logout does make it susceptible
to CSRF attacks, but so does the AWX api logout endpoint, as
things stand today. The most that can happen here is that an
attacker can log someone out. I don't view that as a huge
deal. (Combined with a phishing attack it could be, but
that's a whole separate issue.)

Signed-off-by: Red Hat Contributor <noreply@redhat.com>