AAP-40406 fix gateway SAML legacy migrations (#721)
* AAP-40406 fix gateway SAML legacy migrations

AAP-40406:
With hub using keycloak and controller using SAML with both having identically
named users separate user accounts are created on gateway; this is normal.
Logging into gateway using SAML creates a third account; this, too, is normal.

At this point migration of the legacy accounts is enabled and we logout
from/login to gateway with the expectation that the three accounts are
collapsed into one gateway account.  With the previous could this does
not occur as a consequence of how the accounts get created w/o w/o specific
linkages to their origins.  Because of this the current migration processing
does not automatically remove the accounts.

This change adds a post-individual account migration operation to iterate
over the services and directly remove the account, by ansible_id, from them.

* update comments on old user deletion

link_account() deletes only the local instance of the old user.

* add debug message of service not existing

Using debug level as the current implementation expects this possibility.  Changing the implementation such as to a dynamic discovery of expected services would necessitate increasing the level.

* add debug level message regarding removing the account

* Refactor and add test class

* Remove old comment

* Address code review comments

* Move user creation into fixture

---------

Co-authored-by: Tomas Z <tznamena@redhat.com>