[AAP-48424] Limit collection CI task to devel branch (#836)
* [AAP-48424] Limit collection CI task to devel branch

Based on aap-gateway's development model there should be no
need to run the collection CI against any branch other than devel.

Theoretically there's no reason it couldn't be run on other branches. However,
in preparation for the move to ansible-automation-platform it was observed that
although the CI was spun up it did not actually execute for PRs targeted to
stable-25.

As a practical matter combining the referenced developmental model and the
interest of completing the move to ansible-automation-platform we are disabling
collection CI on anything other than devel.  As time permits we may wish to
revisit this resolve running collection against stable branches.

* Re-trigger CI