AAP-47284 Unified access list and track RBAC across AAP (#901)
requires ansible/django-ansible-base#749

For AAP deployment, all 5 branches for DAB + components is needed. Linking to component PRs:

AAP-48412 Unified RBAC display of permissions, sync to resource server ansible/galaxy_ng#2635
AAP-47283 [2.6] Unified display of RBAC & synchronization ansible/tower#7001
AAP-48412 Unified RBAC display, sync with resource server ansible/eda-server#1363

Also see review doc for this, covering all the patches.

    What is being changed?

This should cover the full scope of ANSTRAT-1464 with 2 exceptions that we do not believe are imminently necessary for functionality and ability to test:

    fixing the JWT header limit problem (not caused by this, but this affects its solution)
    adding logic to sync role assignments in DAB periodic_resource_sync

The gist is: We need to present the user with a coherent display of who-has-permission-to-what. To do that we need to put all permissions to a single thing in one place. Functionally, all permissions get copied over to Gateway. Ideally, Gateway becomes the source-of-truth for RBAC, and we are writing client-facing docs to this effect, but we also do a bidirectional sync in anticipation of migration problems from customer automation and ATF tests themselves.

    Why is this change needed?

RBAC management was very bad in AAP 2.5. We couldn't make it good because it was split over multiple services, although orgs/teams/users were shared, the answer of where permissions live was nuanced and difficult to follow. Very few people could tell you correctly.

    How does this change address the issue?

This brings back an access list that is much of the same style we had in controller in AAP 2.5.