[AAP-49995] Restore migration of role_user_assignments  (#915)
* Restores functionality to migrate role assignments in order to fix it

Revert "Revert "[AAP-47852] Migrate user role assignments from Controller and Hub" (#914)"

This reverts commit e4e1f8c15258c20b6ad2b25cf5fddae20a55c77a.

* Fixes bug reported by AAP-49995

Fetch of role_user_assignments in #900 failed because of an assumption that the path
to `role_user_assignments/` was consistent for both hub and controller relative to the
path of `service-index/`.

This change fixes that assumption, providing the correct relative path for hub and
controller respectively.

Confirmed in aap-dev 2.6 and pytest

* Remove leftover debug logging

* isort

* Update to create DABContentType objects as needed

* Restore previous behavior in launch.py

* Add mock needed for direct call

* Gracefully handle when role migration raises exception

* Reworks to use service-index/role-user-assignments

- This endpoint is consistent across all services so no special cases for the relative URL are needed
- It also provides ansible_id for users and assigned objects, simplifying lookup code.
- Able to delete a lot of code here

* Simplify migrate_user_role_assignments

The _lookup* methods are unnecessary now that ansible_ids are returned in the flat assignments listing

* Update to use client.list_user_assignments

Uses new methods in https://github.com/ansible/django-ansible-base/pull/786 instead of calling private client._make_request

* Tests to cover exception cases

* Trigger build

* Constrain setuptools_scm due to issues with newly-released 9.0.x