AAP-48835 Configure proxied X-FORWARDED-PROTO in django and envoy (#929)
fix: configure proxied X-FORWARDED-PROTO in django and envoy

This sets the SECURE_PROXY_SSL_HEADER in defaults.py to X-FORWARDED-PROTO

Envoy drops this header if we don't set xff-num-trusted-hops to the correct
number based on the number of hops between envoy the the SSL terminating proxy.

The result of this this is that we still need a setting (because we have no way
of knowing how many proxies sit in front of the gateway).

AAP-48835